Defer building types hash until it's needed
