+++ lib/graphql/static_validation/rules/fragment_spreads_are_possible.rb
@@ -45,9 +45,10 @@ module GraphQL
           return
         end
         parent_types = context.warden.possible_types(parent_type.unwrap)
-        child_types = context.warden.possible_types(child_type.unwrap)
-
-        if child_types.none? { |c| parent_types.include?(c) }
+        if !parent_types.include?(child_type) && (
+            (child_types = context.warden.possible_types(child_type.unwrap)) &&
+            child_types.none? { |c| parent_types.include?(c) }
+          )
           name = node.respond_to?(:name) ? " #{node.name}" : ""
           add_error(GraphQL::StaticValidation::FragmentSpreadsArePossibleError.new(
             "Fragment#{name} on #{child_type.graphql_name} can't be spread inside #{parent_type.graphql_name}",
